Solution to Task 2
Solves task 2.

The method `exercise2` is the improved version of `exercise1`. This
version takes into account the possibility of a team being a member of
another team.

The list comprehension solution from task 1 is replace with a lambda function,
which with the help of the closure `is_in_team`, it checks if the person is
part of the team.

`is_in_team` is used recursively to improve performace.
While it works for data2, it is prone to infinite loops.

Add unit test for method `exercise2`.